Use classmethods to provide convenience methods for making KubeClusters
